+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [Unreleased]
+- fix global shortcuts recording
 
 ## [1.0.0] - 2026-04-02
 ### Added